Correction of a bug and notification improvement
Bug correction: I had forgotten to shutdown one ScheduledExecutorService, 
so BlueGPS was always trying to reconnect to the GPS, 
even after the application and service had been stopped.

Improvement: 
First, when the connection with the GPS is lost, 
a notification is displayed.
And then, if  the GPS connection service is closed because of two many connection retry, an second notification is displayed.
The fist notification allows the user to close BlueGPS service and the second one allows the user to re-enable it.